requirements.txt
celery==4.3.0
Flask==1.1.1
Flask-WTF==0.14.2
redis==3.3.11
requests==2.22.0
SQLAlchemy==1.3.10
Flask-SQLAlchemy==2.4.1
pyjwt